\docType{package}
\name{activeH5}
\alias{activeH5}
\alias{activeH5-package}
\title{activeH5 is a tool for big data objects in R on disk and in memory created by Active Analytics Ltd.}
\description{
  activeH5 provides a set of tools to allow large data
  frames and matrices to be written to file and to memory
  and are stored in chunks. On file the data is stored in
  H5 format and in memory pointers are used to access
  chunks of data. The data objects are written as reference
  classes and are h5DF, h5MAT, h5MEMDF, and h5MEMMAT.
}
\details{
  The h5DF object is used to store very large data frames
  on disk. The data itself is stores as purely numerical
  data. Character columns are converted to factors and
  factors to numeric data, then each chunk is stores as a
  matrix. This means that the I/O speed is fast. Meta data
  associated with factors are stored in the file and the
  h5DF object and are recombined with the daat as it is
  read back into R.

  The h5MAT object is used to store very large matrices on
  disk in chunks. The data is assumed to be numeric data so
  I/O speed is fast even faster than h5DF.

  The h5MEMDF object is used to store very large data
  frames in memory as chunks with pointers to each chunk.
  Data can only be stored where there is sufficient memory
  to accomodate the object(s). Storing data like this makes
  it easier move, manipulate and work with very large
  datasets without the constant replication that occurs
  when using them in normal R functions. Since the data is
  stored in memory access time is faster than when the data
  is stored on disk.

  The h5MEMMAT object is used to store very large matrices
  in memory as chunks. Again there are pointers to each
  chunk of data and there has to be sufficient memory to
  accomodate the object(s). Storing data like this makes it
  easier move, manipulate and work with very large datasets
  without the constant replication that occurs when using
  them in normal R functions. Since the data is stored in
  memory access time is faster than when stored on disk.

  The HDF5 files generated by h5DF, h5MAT are not designed
  to be used with other software that use HDF5 objects. Any
  data that is created using the activeH5 module may not
  read back as expected with other HDF5 APIs.
}

