\name{h5ChunkSel}
\alias{h5ChunkSel}
\title{Function returns a column subset of data frame chunk as a list}
\usage{
  h5ChunkSel(chunkName, selCols, filePath)
}
\arguments{
  \item{chunkName}{the name of the chunk to be read}

  \item{selCols}{the columns that will be selected}

  \item{filePath}{the path to the h5 file}
}
\value{
  List representing chunk data frame
}
\description{
  Function returns a column subset of data frame chunk as a
  list
}

