\name{h5CreateMetaData}
\alias{h5CreateMetaData}
\title{Function creates the groups for the meta data to be written to the h5 file}
\usage{
  h5CreateMetaData(filePath)
}
\arguments{
  \item{filePath}{character path to the location where the
  h5 file will be written}
}
\value{
  int 0
}
\description{
  Function to create the groups in the h5 file before it is
  populated. This function is intended for internal use
  only
}

