\name{h5FlushFile}
\alias{h5FlushFile}
\title{Function to finalize h5 file contents allowing them to be read by another user}
\usage{
  h5FlushFile(filePath)
}
\arguments{
  \item{filePath}{character path to the file which will be
  flushed}
}
\value{
  int 0
}
\description{
  The same as the h5CloseFile() function
}

