\name{h5ReadCharVector}
\alias{h5ReadCharVector}
\title{Function to read a character vector from meta data}
\usage{
  h5ReadCharVector(charName, filePath)
}
\arguments{
  \item{charName}{character the name of the meta data item
  to be read back from file}

  \item{filePath}{character for the path to the file where
  the item will be read}
}
\value{
  character containing the character vector that has been
  read from meta data
}
\description{
  Function to read a character vector from meta data
}

