\name{h5ReadDoubleMat}
\alias{h5ReadDoubleMat}
\title{Function to read a matrix chunk from a h5 file}
\usage{
  h5ReadDoubleMat(chunkName, filePath)
}
\arguments{
  \item{chunkName}{the name of the chunk to be read back}

  \item{filePath}{the path to the h5 file}
}
\value{
  matrix chunk defined by chunkName
}
\description{
  Function to read a matrix chunk from a h5 file
}

