\name{h5ReadDoubleMat3}
\alias{h5ReadDoubleMat3}
\title{Legacy function to return a data frame chunk as a list}
\usage{
  h5ReadDoubleMat3(chunkName, filePath)
}
\arguments{
  \item{chunkName}{the name of the chunk to be read}

  \item{filePath}{the path to the h5 file}
}
\value{
  List of the data frame chunk
}
\description{
  Experimental function not intended for use at all
}

