\name{h5ReadFactor}
\alias{h5ReadFactor}
\title{Function to read factor levels from meta data for a pgiven factor}
\usage{
  h5ReadFactor(charName, filePath)
}
\arguments{
  \item{charName}{character denoting the factor from the
  meta data factor}

  \item{filePath}{character denoting the path to the h5
  file}
}
\value{
  character of factor levels
}
\description{
  Function to read factor levels from meta data for a
  pgiven factor
}

