\name{h5ReadInt}
\alias{h5ReadInt}
\title{Function to read an integer item from meta data}
\usage{
  h5ReadInt(intName, filePath)
}
\arguments{
  \item{intName}{character for the name of the item to be
  read back}

  \item{filePath}{character for the path to the h5 file}
}
\value{
  int iteger item defined by intName in the meta data
}
\description{
  Function to read an integer item from meta data
}

