\name{h5ReadMemMAT}
\alias{h5ReadMemMAT}
\title{Function to read a matrix from memory when given the external pointer}
\usage{
  h5ReadMemMAT(ptr)
}
\arguments{
  \item{ptr}{an externalptr to the matrix object to be
  retrieved}
}
\value{
  matrix that is held at the externalptr
}
\description{
  Intended for internal use only
}

